refactor: View를 InputView와 OuputView로 분리
